refactor: move JwtAuthenticationFilter to filters package
